+++ hyperspy/external/astroML/LICENSE.RST
