+++ hyperspy/external/astroML/bayesian_blocks.py
