+++ hyperspy/external/tifffile.py
