+++ hyperspy/io_plugins/tiff.py
@@ -26,7 +26,7 @@ try:
 except ImportError:
     with warnings.catch_warnings():
         warnings.simplefilter("ignore")
-        from hyperspy.misc.borrowed.tifffile import imsave, TiffFile
+        from hyperspy.external.tifffile import imsave, TiffFile
     warnings.warn(
         "Failed to import the optional scikit image package. "
         "Loading of some compressed images will be slow.\n")
