+++ setup.py
@@ -163,8 +163,8 @@ with update_version_when_dev() as version:
                   'hyperspy.misc.machine_learning',
                   'hyperspy.misc.mpfit',
                   'hyperspy.misc.mpfit.tests',
-                  'hyperspy.misc.borrowed',
-                  'hyperspy.misc.borrowed.astroML',
+                  'hyperspy.external',
+                  'hyperspy.external.astroML',
                   ],
         requires=install_req,
         scripts=scripts,
